+++ theme/components/SakuraPostList.vue
@@ -2,6 +2,8 @@
 import { computed } from 'vue'
 import { useSiteConfig, useSiteStore } from 'valaxy'
 import type { Post } from 'valaxy'
+import { useThemeConfig } from '../composables'
+import { useSakuraAppStore } from '../stores/app'
 
 // import { usePostList } from 'valaxy'
 
@@ -15,10 +17,13 @@ const props = withDefaults(defineProps<{
 
 const site = useSiteStore()
 const siteConfig = useSiteConfig()
+const themeConfig = useThemeConfig()
+const sakura = useSakuraAppStore()
 const pageSize = computed(() => siteConfig.value.pageSize)
 
 // const routes = usePostList({ type: props.type || '' })
 // const posts = computed(() => props.posts || routes.value)
+const paginationType = computed(() => themeConfig.value.pagination?.type || 'infinite-scroll')
 const posts = computed(() => (
   props.posts || site.postList).filter(post => import.meta.env.DEV ? true : !post.hide),
 )
@@ -26,7 +31,7 @@ const posts = computed(() => (
 const displayedPosts = computed(() =>
   posts.value.slice(
     (props.curPage - 1) * pageSize.value,
-    props.curPage * pageSize.value,
+    props.curPage * pageSize.value * sakura.loadMultiple,
   ),
 )
 
@@ -38,6 +43,10 @@ const postsWithLimitedTags = computed(() => {
     return post
   })
 })
+
+function handLoadMore() {
+  sakura.handLoadMore()
+}
 </script>
 
 <template>
@@ -53,6 +62,12 @@ const postsWithLimitedTags = computed(() => {
       </Transition>
     </template>
 
-    <ValaxyPagination :cur-page="curPage" :page-size="pageSize" :total="posts.length" />
+    <template v-if="paginationType === 'infinite-scroll'">
+      <SakuraPagination :cur-page="curPage" :page-size="pageSize" :total="posts.length" @load-more="handLoadMore" />
+    </template>
+
+    <template v-if="paginationType === 'pagination'">
+      <ValaxyPagination :cur-page="curPage" :page-size="pageSize" :total="posts.length" />
+    </template>
   </div>
 </template>